# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sigma Scholar API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ClassItem {
  id: string;
  name: string;
  order: number;
  /** @nullable */
  totalSubjects?: number | null;
  /** @nullable */
  totalChapters?: number | null;
}

export interface Subject {
  id: string;
  classId: string;
  name: string;
  color: string;
  icon: string;
  order?: number;
  /** @nullable */
  totalChapters?: number | null;
}

export interface Chapter {
  id: string;
  subjectId: string;
  classId: string;
  name: string;
  order: number;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  totalTopics?: number | null;
  /** @nullable */
  hasNotes?: boolean | null;
  /** @nullable */
  hasPDF?: boolean | null;
}

export interface Topic {
  id: string;
  chapterId: string;
  name: string;
  order: number;
  /** @nullable */
  description?: string | null;
}

export type NoteType = typeof NoteType[keyof typeof NoteType];


export const NoteType = {
  pdf: 'pdf',
  text: 'text',
  video: 'video',
} as const;

export interface Note {
  id: string;
  classId: string;
  subjectId: string;
  chapterId: string;
  title: string;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  pdfUrl?: string | null;
  type: NoteType;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export type NoteInputType = typeof NoteInputType[keyof typeof NoteInputType];


export const NoteInputType = {
  pdf: 'pdf',
  text: 'text',
  video: 'video',
} as const;

export interface NoteInput {
  classId: string;
  subjectId: string;
  chapterId: string;
  title: string;
  content?: string;
  pdfUrl?: string;
  type: NoteInputType;
}

export type NoteUpdateType = typeof NoteUpdateType[keyof typeof NoteUpdateType];


export const NoteUpdateType = {
  pdf: 'pdf',
  text: 'text',
  video: 'video',
} as const;

export interface NoteUpdate {
  title?: string;
  content?: string;
  pdfUrl?: string;
  type?: NoteUpdateType;
}

export type TestType = typeof TestType[keyof typeof TestType];


export const TestType = {
  chapter: 'chapter',
  subject: 'subject',
  mock: 'mock',
  exam: 'exam',
} as const;

export interface Test {
  id: string;
  classId: string;
  subjectId: string;
  /** @nullable */
  chapterId?: string | null;
  title: string;
  type: TestType;
  duration: number;
  totalMarks: number;
  totalQuestions?: number;
  createdAt: string;
}

export type TestInputType = typeof TestInputType[keyof typeof TestInputType];


export const TestInputType = {
  chapter: 'chapter',
  subject: 'subject',
  mock: 'mock',
  exam: 'exam',
} as const;

export interface TestInput {
  classId: string;
  subjectId: string;
  chapterId?: string;
  title: string;
  type: TestInputType;
  duration: number;
  totalMarks: number;
  questions?: string;
}

export interface TestUpdate {
  title?: string;
  duration?: number;
  totalMarks?: number;
  questions?: string;
}

export interface TestResultInput {
  userId: string;
  score: number;
  totalMarks: number;
  timeTaken: number;
  answers?: string;
}

export interface TestResult {
  id: string;
  testId: string;
  userId: string;
  score: number;
  totalMarks: number;
  percentage: number;
  timeTaken?: number;
  createdAt: string;
}

export interface UserProfile {
  id: string;
  email: string;
  displayName: string;
  /** @nullable */
  photoURL?: string | null;
  /** @nullable */
  classId?: string | null;
  /** @nullable */
  className?: string | null;
  createdAt: string;
  /** @nullable */
  lastActive?: string | null;
  /** @nullable */
  isAdmin?: boolean | null;
}

export interface UserUpdate {
  displayName?: string;
  classId?: string;
  photoURL?: string;
}

export interface ProgressEntry {
  id: string;
  userId: string;
  chapterId: string;
  subjectId: string;
  classId: string;
  completed: boolean;
  /** @nullable */
  percentage?: number | null;
  updatedAt: string;
}

export interface ProgressInput {
  userId: string;
  chapterId: string;
  subjectId: string;
  classId: string;
  completed: boolean;
  percentage?: number;
}

export type BookmarkType = typeof BookmarkType[keyof typeof BookmarkType];


export const BookmarkType = {
  chapter: 'chapter',
  note: 'note',
  test: 'test',
} as const;

export interface Bookmark {
  id: string;
  userId: string;
  type: BookmarkType;
  refId: string;
  title: string;
  /** @nullable */
  subtitle?: string | null;
  createdAt: string;
}

export type BookmarkInputType = typeof BookmarkInputType[keyof typeof BookmarkInputType];


export const BookmarkInputType = {
  chapter: 'chapter',
  note: 'note',
  test: 'test',
} as const;

export interface BookmarkInput {
  userId: string;
  type: BookmarkInputType;
  refId: string;
  title: string;
  subtitle?: string;
}

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ChatMessage {
  id: string;
  userId: string;
  role: ChatMessageRole;
  content: string;
  createdAt: string;
}

export type ChatMessageInputRole = typeof ChatMessageInputRole[keyof typeof ChatMessageInputRole];


export const ChatMessageInputRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ChatMessageInput {
  userId: string;
  role: ChatMessageInputRole;
  content: string;
}

export interface ActivityItem {
  id: string;
  type: string;
  description: string;
  /** @nullable */
  userId?: string | null;
  /** @nullable */
  userName?: string | null;
  createdAt: string;
}

export interface DashboardStats {
  totalStudents: number;
  totalNotes: number;
  totalTests: number;
  totalChapters: number;
  recentActivity: ActivityItem[];
}

export interface UserStats {
  userId: string;
  completedChapters: number;
  totalChapters: number;
  testsAttempted: number;
  averageScore: number;
  studyStreak: number;
  recentResults?: TestResult[];
}

export type ListNotesParams = {
classId?: string;
subjectId?: string;
chapterId?: string;
};

export type ListTestsParams = {
classId?: string;
subjectId?: string;
type?: string;
};

export type ListResultsParams = {
userId?: string;
};