# lib/api-zod/src/generated/types/noteUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sigma Scholar API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NoteUpdateType } from './noteUpdateType';

export interface NoteUpdate {
  title?: string;
  content?: string;
  pdfUrl?: string;
  type?: NoteUpdateType;
}