# lib/api-zod/src/generated/types/progressEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sigma Scholar API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProgressEntry {
  id: string;
  userId: string;
  chapterId: string;
  subjectId: string;
  classId: string;
  completed: boolean;
  /** @nullable */
  percentage?: number | null;
  updatedAt: string;
}