# lib/api-zod/src/generated/types/testInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sigma Scholar API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TestInputType } from './testInputType';

export interface TestInput {
  classId: string;
  subjectId: string;
  chapterId?: string;
  title: string;
  type: TestInputType;
  duration: number;
  totalMarks: number;
  questions?: string;
}