# lib/api-zod/src/generated/types/testUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sigma Scholar API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TestUpdate {
  title?: string;
  duration?: number;
  totalMarks?: number;
  questions?: string;
}